refactor(request): extract approval comment rendering helpers
